split out timepicker redux-form tests/interactor
