+++ lib/Timepicker/tests/Timepicker-test.js
@@ -1,13 +1,8 @@
 import React from 'react';
 import { describe, beforeEach, it } from 'mocha';
 import moment from 'moment-timezone';
-import sinon from 'sinon';
-import { Field } from 'redux-form';
 import {
   runAxeTest,
-  HTML,
-  TextField,
-  IconButton,
   Button,
   including,
   converge,
@@ -17,62 +12,13 @@ import {
 import translationsDE from '../../../translations/stripes-components/de';
 
 import { mountWithContext } from '../../../tests/helpers';
-import TestForm from '../../../tests/TestForm';
-
 import Timepicker from '../Timepicker';
 import TimepickerHarness from './TimepickerHarness';
+import {
+  Timepicker as TimepickerInteractor,
+  TimeDropdown as TimepickerDropdownInteractor
+} from './timepicker-interactor-bt';
 
-const TimepickerDropdownInteractor = HTML.extend('time dropdown')
-  .selector('[data-test-timepicker-dropdown]')
-  .filters({
-    periodToggle: Button({ id: including('period-toggle') }).exists(),
-    hour: (el) => {
-      const node = el.querySelector('input[data-test-timepicker-dropdown-hours-input]');
-      return node.value ? parseInt(node.value, 10) : '';
-    },
-    minute: (el) => {
-      const node = el.querySelector('input[data-test-timepicker-dropdown-minutes-input]');
-      return node.value ? parseInt(node.value, 10) : '';
-    }
-  })
-  .actions({
-    clickPeriodToggle: ({ find }) => find(Button({ id: including('period-toggle') })).click(),
-    fillHour: ({ find }, value) => find(TextField({ id: including('hour-input') }))
-      .perform((el) => {
-        const node = el.querySelector('input');
-        const property = Object.getOwnPropertyDescriptor(Object.getPrototypeOf(node), 'value');
-        property.set.call(node, value);
-        node.dispatchEvent(new InputEvent('input', { inputType: 'insertFromPaste', bubbles: true, cancelable: false }));
-      }),
-    fillMinute: ({ find }, value) => find(TextField({ id: including('minute-input') }))
-      .perform((el) => {
-        const node = el.querySelector('input');
-        const property = Object.getOwnPropertyDescriptor(Object.getPrototypeOf(node), 'value');
-        property.set.call(node, value);
-        node.dispatchEvent(new InputEvent('input', { inputType: 'insertFromPaste', bubbles: true, cancelable: false }));
-      }),
-    clickConfirm: ({ find }) => find(Button({ id: including('set-time') })).click(),
-    clickAddMinute: ({ find }) => find(IconButton({ id: including('next-minute') })).click(),
-    clickAddHour: ({ find }) => find(IconButton({ id: including('next-hour') })).click(),
-    clickDecrementMinute: ({ find }) => find(IconButton({ id: including('prev-minute') })).click(),
-    clickDecrementHour: ({ find }) => find(IconButton({ id: including('prev-hour') })).click(),
-  });
-
-const TimepickerInteractor = TextField.extend('timepicker')
-  .actions({
-    clickDropdownToggle: ({ find }) => find(IconButton({ icon: 'clock' })).click(),
-    clickInput: ({ perform }) => perform((el) => el.querySelector('input').click()),
-    focusDropdownButton: ({ find }) => find(IconButton({ icon: 'clock' })).focus(),
-    clickClear: ({ find }) => find(IconButton({ icon: 'times-circle-solid' })).click(),
-    fillInput: ({ find }, value) => find(TextField())
-      .perform((el) => {
-        const node = el.querySelector('input');
-        const property = Object.getOwnPropertyDescriptor(Object.getPrototypeOf(node), 'value');
-        property.set.call(node, value);
-        node.dispatchEvent(new InputEvent('input', { inputType: 'insertFromPaste', bubbles: true, cancelable: false }));
-      }),
-    blur: () => document.body.click(),
-  });
 
 describe('Timepicker', () => {
   const timepicker = TimepickerInteractor();
@@ -249,44 +195,6 @@ describe('Timepicker', () => {
     });
 
     it('emits an event with the time formatted as displayed', () => converge(() => timeOutput === '05:00 PM'));
-  });
-
-  describe('coupled to redux form', () => {
-    describe('selecting a time', () => {
-      let timeOutput;
-      const handleChange = sinon.spy((e) => {
-        timeOutput = e.target.value;
-      });
-      beforeEach(async () => {
-        timeOutput = '';
-        handleChange.resetHistory();
-        await mountWithContext(
-          <TestForm>
-            <Field
-              component={
-                ({ input, meta }) => {
-                  input.onChange = (e) => {
-                    input.onChange(e);
-                    handleChange(e);
-                  };
-                  return (
-                    <Timepicker
-                      input={input}
-                      meta={meta}
-                    />
-                  )
-                }}
-              name="time"
-            />
-          </TestForm>
-        );
-
-        await timepicker.fillIn('05:00 PM');
-      });
-
-      it('returns an ISO 8601 time string at UTC', () => converge(() => timeOutput === '17:00:00.000Z'));
-      it('calls the handler multiple times - on change and onBlur', () => converge(() => handleChange.calledTwice));
-    });
   });
 
   describe('selecting a time with timeZone prop', () => {
